template nix : fix aws

+++ templates/nix_shell/nixfiles/components/aws.nix
@@ -3,10 +3,11 @@
 {
   buildInputs = (with pkgs; [ python3 awscli2 ]);
     
+  # awsebcli package marked as broken in NixOS 21.05, we install it manually
   shellInit = ''
     python3 -m venv .venv
     source .venv/bin/activate
-    pip install awsebcli
+    test -e .venv/bin/eb || pip install awsebcli
     '';
 
   shellStartService = ''
